Add the `LanguagesFlashCard` object
This is an object used for languages courses (e.g. French vocabulary) to include additional functionality to the basic flashcard (the `FlashCard` interface). The extra functionality includes the difficulty level (tier) of the vocabulary (i.e. General vocabulary, or Foundation or Higher tier) in addition to an answer 'prefix' - data for this could be something such as 'le' or 'la' in some French vocabulary.

 Updates #13