Add option to choose content type for viewing flashcards
The layout XML file (`activity_topic_detail.xml`) has been changed to
include an additional spinner so that users can choose the content type
for viewing flashcards - for 'standard' flashcards, this is viewing
Paper 1, Paper 2, or all content; and for 'languages' flashcards, it
would be `General vocabulary`, `Foundation tier`, and `Higher tier`.
This explains that `spinner_items.xml` includes the string arrays that
could be displayed on this additional spinner; and the label for this
spinner has been added to `strings.xml`.

The code for accessing the content type values from the CSV files was
added in previous commits.

There are more changes made the Java files (from this commit). The
StandardFlashCard object now stores an `isPaper2` boolean and parts of
`Topic.java` have been modified for this. In addition, `Topic.java`
contains a new static inner class which contains two static methods for
filtering 'standard' and 'languages' flashcards, depending on what the
user chose from the spinner (i.e. a parameter of the methods).

Between the TopicDetailActivity and FlashCardActivity, there is a new
Intent to pass the list of generated flashcards that the user will view,
as opposed to generating this list in FlashCardActivity. The
TopicDetailActivity has also been modified to deal with user input for
the content of the flashcards they will view (from the additional
spinner).

Resolves #15