Access data via CSVs directly where SQLite was previously used
To simplify the loading process, and as there is not a lot of data (as in not thousands of lines long), I have changed some files (including deleting and adding new ones) so that the data is accessed from the CSV file(s) directly rather than being accessed via SQLite database(s) which would be generated from these CSV files when the app loads for the first time.

This also removes the need for an `InitializeDbActivity` or the need for SQLiteHelper classes. I have created `MainDatabases.java` to provide static methods for accessing data from the CSVs directly.