Layout optimizations for topic detail ui
The two new files: `topic_detail_header.xml` and `topic_detail_options.xml` contain specific parts of the topic detail UI which are used in the two `activity_topic_detail.xml` layout files (for different screen densities).

`TopicDetailActivity.java` needed to be updated so that the options layout is cast to a ViewGroup rather than a LinearLayout.